🖼️🎨 Extract icons and colors logic from `CategoryCreationHelper` to responsible picker dialogs itself
