+++ app/src/main/java/com/example/sunrise/utils/ColorPickerDialog.java
@@ -1,6 +1,7 @@
 package com.example.sunrise.utils;
 
 import android.content.Context;
+import android.graphics.Color;
 import android.view.LayoutInflater;
 import android.view.View;
 
@@ -12,7 +13,9 @@ import com.example.sunrise.R;
 import com.example.sunrise.adapters.ColorsAdapter;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 
 public class ColorPickerDialog {
     private final AlertDialog dialog;
@@ -23,8 +26,9 @@ public class ColorPickerDialog {
         void onColorSelected(int color);
     }
 
-    public ColorPickerDialog(Context context, List<Integer> colors, OnColorSelectedListener listener) {
-        this.colors = colors;
+    public ColorPickerDialog(Context context, OnColorSelectedListener listener) {
+        // Initialize colors
+        this.colors = initializeColors();
         this.colorSelectedListener = listener;
 
         View dialogView = LayoutInflater.from(context).inflate(R.layout.dialog_choose_color, null);
@@ -61,5 +65,32 @@ public class ColorPickerDialog {
         if (dialog != null && dialog.isShowing()) {
             dialog.dismiss();
         }
+    }
+
+    private List<Integer> initializeColors() {
+        List<Integer> colors = new ArrayList<>();
+
+        // Define hexadecimal colors
+        String[] hexValues = {
+                "#FFCCCC", // Pastel Red
+                "#FFE5CC", // Pastel Orange
+                "#FFF2CC", // Pastel Yellow
+                "#CCFFCC", // Pastel Green
+                "#CCE5FF", // Pastel Blue
+                "#FFCCFF"  // Pastel Purple
+        };
+
+        // Convert hexadecimal values to color integers and add them to the list
+        for (String hex : hexValues) {
+            int color = Color.parseColor(hex);
+            colors.add(color);
+        }
+
+        return colors;
+    }
+
+    public int getRandomColor() {
+        Random random = new Random();
+        return colors.get(random.nextInt(colors.size()));
     }
 }
